+++ dalesbred/src/main/java/org/dalesbred/internal/instantiation/PropertyAccessor.java
@@ -25,11 +25,14 @@ package org.dalesbred.internal.instantiation;
 import org.dalesbred.annotation.DalesbredIgnore;
 import org.dalesbred.internal.utils.Throwables;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 import java.lang.reflect.Field;
+import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.lang.reflect.Type;
 import java.util.Optional;
+import java.util.regex.Pattern;
 
 import static java.lang.reflect.Modifier.isPublic;
 import static org.dalesbred.internal.utils.StringUtils.isEqualIgnoringCaseAndUnderscores;
@@ -41,13 +44,48 @@ abstract class PropertyAccessor {
     abstract Type getType();
 
     @NotNull
-    static Optional<PropertyAccessor> findAccessor(@NotNull Class<?> cl, @NotNull String name) {
-        Optional<PropertyAccessor> setter = findSetter(cl, name).map(SetterPropertyAccessor::new);
+    private static final Pattern PERIOD = Pattern.compile("\\.");
+
+    @NotNull
+    static Optional<PropertyAccessor> findAccessor(@NotNull Class<?> cl, @NotNull String path) {
+        String[] segments = PERIOD.split(path, -1);
+
+        assert segments.length != 0; // split will always return non-empty array
+
+        if (segments.length == 1)
+            return findFinalAccessor(cl, segments[0]);
+
+        Class<?> currentClass = cl;
+
+        PropertyReader[] readers = new PropertyReader[segments.length - 1];
+        for (int i = 0; i < segments.length - 1; i++) {
+            Field field = findField(currentClass, segments[i]).orElse(null);
+            if (field != null) {
+                readers[i] = field::get;
+                currentClass = field.getType();
+            } else {
+                Method getter = findGetter(currentClass, segments[i]).orElse(null);
+                if (getter != null) {
+                    readers[i] = getter::invoke;
+                    currentClass = getter.getReturnType();
+                } else {
+                    return Optional.empty();
+                }
+            }
+        }
+
+        Optional<PropertyAccessor> accessor = findFinalAccessor(currentClass, segments[segments.length - 1]);
+        return accessor.map(a -> new NestedPathAccessor(readers, path, a));
+    }
+
+    @NotNull
+    private static Optional<PropertyAccessor> findFinalAccessor(@NotNull Class<?> currentClass, @NotNull String name) {
+        Optional<PropertyAccessor> setter = findSetter(currentClass, name).map(SetterPropertyAccessor::new);
 
         if (setter.isPresent()) {
             return setter;
         } else {
-            return findField(cl, name).map(FieldPropertyAccessor::new);
+            return findField(currentClass, name).map(FieldPropertyAccessor::new);
         }
     }
 
@@ -67,14 +105,27 @@ abstract class PropertyAccessor {
 
     @NotNull
     private static Optional<Method> findSetter(@NotNull Class<?> cl, @NotNull String name) {
+        return findGetterOrSetter(cl, name, false);
+    }
+
+    @NotNull
+    private static Optional<Method> findGetter(@NotNull Class<?> cl, @NotNull String name) {
+        return findGetterOrSetter(cl, name, true);
+    }
+
+    @NotNull
+    private static Optional<Method> findGetterOrSetter(@NotNull Class<?> cl, @NotNull String propertyName, boolean getter) {
+        String methodName = (getter ? "get" : "set") + propertyName;
+        int parameterCount = getter ? 0 : 1;
         Method result = null;
 
-        String methodName = "set" + name;
         for (Method method : cl.getMethods()) {
-
-            if (isPublic(method.getModifiers()) && isEqualIgnoringCaseAndUnderscores(methodName, method.getName()) && method.getParameterTypes().length == 1 && !method.isAnnotationPresent(DalesbredIgnore.class)) {
+            if (isPublic(method.getModifiers())
+                    && isEqualIgnoringCaseAndUnderscores(methodName, method.getName())
+                    && method.getParameterCount() == parameterCount
+                    && !method.isAnnotationPresent(DalesbredIgnore.class)) {
                 if (result != null)
-                    throw new InstantiationFailureException("Conflicting setters for property: " + result + " - " + name);
+                    throw new InstantiationFailureException("Conflicting accessors for property: " + result + " - " + propertyName);
                 result = method;
             }
         }
@@ -133,6 +184,60 @@ abstract class PropertyAccessor {
                 throw Throwables.propagate(e);
             }
         }
+    }
+
+    private static final class NestedPathAccessor extends PropertyAccessor {
+
+        @NotNull
+        private final PropertyReader[] readers;
+
+        @NotNull
+        private final String path;
+
+        @NotNull
+        private final PropertyAccessor accessor;
+
+        public NestedPathAccessor(@NotNull PropertyReader[] readers, @NotNull String path, @NotNull PropertyAccessor accessor) {
+            this.readers = readers;
+            this.path = path;
+            this.accessor = accessor;
+        }
+
+        @Override
+        Type getType() {
+            return accessor.getType();
+        }
+
+        @Override
+        void set(Object object, Object value) {
+            accessor.set(resolveFinalObject(object), value);
+        }
+
+        @NotNull
+        private Object resolveFinalObject(@NotNull Object object) {
+            try {
+                Object obj = object;
+
+                for (PropertyReader reader : readers) {
+                    Object value = reader.propertyValue(obj);
+                    if (value != null)
+                        obj = value;
+                    else
+                        throw new InstantiationFailureException(
+                                "Failed to set property for '" + path + "', because one of the intermediate objects was null.");
+                }
+
+                return obj;
+
+            } catch (InvocationTargetException | IllegalAccessException e) {
+                throw new InstantiationFailureException("Failed to set property for '" + path + "'.", e);
+            }
+        }
+    }
+
+    private interface PropertyReader {
+        @Nullable
+        Object propertyValue(@NotNull Object o) throws IllegalAccessException, InvocationTargetException;
     }
 }