+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     ([#19](https://github.com/EvidentSolutions/dalesbred/issues/19))
   - Support binding arrays on Oracle, which does not support standard JDBC API.
   - Add dialect for H2. ([#20](https://github.com/EvidentSolutions/dalesbred/issues/20))
+  - Support property paths when binding (instead of just simple names). ([#7](https://github.com/EvidentSolutions/dalesbred/issues/7))
 
 ## 1.0.0 (2015-05-29)